# swagger.yaml
openapi: 3.0.3
info:
  title: Werfie API
  description: |
    Complete API documentation for Werfie (X-Clone) platform.
    - **Client Gateway**: http://localhost:3001
    - **Admin Backend**: http://localhost:3012
    - **Messaging Service**: http://localhost:3019 (direct or via gateway)

    ## Quick test (always 200)
    **GET /api/health** - No auth. Returns 200 when gateway is up.

    ## Auth Flow (to get 200 on protected endpoints)
    1. Call **POST /api/auth/login** with `{"email":"user1@xclone.com","password":"password123"}`
    2. Copy the `accessToken` from the response (from `data.accessToken`)
    3. Click **Authorize** (lock icon, top right), paste ONLY the token (no "Bearer " prefix)
    4. Protected endpoints (/api/auth/me, /api/auth/logout, /api/auth/change-password) will now return 200

    ## Proxied endpoints (Users, Posts, etc.)
    Endpoints like /api/users/*, /api/posts/* are proxied to microservices. Ensure those services are running (see ecosystem.config.js) or you may get 5xx.
  version: 1.0.0
servers:
  - url: /
    description: Same origin (use this - no CORS)
  - url: http://localhost:3001
    description: Client Gateway (Auth, Users, Posts, etc.)
  - url: http://127.0.0.1:3001
    description: Client Gateway (127.0.0.1)
  - url: http://localhost:3019
    description: Messaging Service (direct)
  - url: http://localhost:3012
    description: Admin Backend
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Paste the accessToken from /api/auth/login (no "Bearer " prefix)
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
...
paths:
...
  /api/spaces:
    get:
      summary: List spaces
      tags: [Spaces]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Active spaces }
        '401': { description: Unauthorized - use Authorize with accessToken from login }
    post:
      summary: Create space
      description: Requires Bearer token. Use Authorize with accessToken from POST /api/auth/login.
      tags: [Spaces]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                scheduledAt: { type: string, format: date-time }
      responses:
        '201': { description: Space created }
        '401': { description: Unauthorized - use Authorize with accessToken from login }
  /api/spaces/{id}/start:
    post:
      summary: Start space
      description: Requires Bearer token. Use Authorize with accessToken from login.
      tags: [Spaces]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Space started }
        '401': { description: Unauthorized - use Authorize with accessToken from login }
  /api/spaces/{id}/end:
    post:
      summary: End space
      description: Requires Bearer token. Use Authorize with accessToken from login.
      tags: [Spaces]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Space ended }
        '401': { description: Unauthorized - use Authorize with accessToken from login }

  # ============ LISTS ============
  # All list endpoints require Bearer token. Use Authorize with accessToken from POST /api/auth/login.
...
  /api/comments/create:
    post:
      summary: Create comment
      tags: [Comments]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                announcementId: { type: string }
                content: { type: string }
                parentCommentId: { type: string }
      responses:
        '201': { description: Created }
  /api/comments/{announcementId}:
    get:
      summary: Get comments for announcement
      tags: [Comments]
      parameters:
        - name: announcementId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Comments }
  /api/comments/report:
    post:
      summary: Report comment
      tags: [Comments]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                commentId: { type: string }
                reason: { type: string }
      responses:
        '200': { description: Reported }
  /api/comments/moderation/queue:
    get:
      summary: Get moderation queue
      tags: [Comments]
      parameters:
        - name: status
          in: query
          schema: { type: string }
      responses:
        '200': { description: Queue }
  /api/comments/moderation/approve/{queueId}:
    post:
      summary: Approve comment
      tags: [Comments]
      parameters:
        - name: queueId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Approved }
  /api/comments/moderation/reject/{queueId}:
    post:
      summary: Reject comment
      tags: [Comments]
      parameters:
        - name: queueId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Rejected }
  /api/comments/moderation/fact-flag/{commentId}:
    post:
      summary: Fact-flag comment
      tags: [Comments]
      parameters:
        - name: commentId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Flagged }

  # ============ CRISIS ============
...